+++ src/components/input/Input.stories.ts
@@ -6,6 +6,19 @@ import WvuiButton from '../button/Button.vue';
 import { InputType } from './InputType';
 import './Input.stories.less';
 
+const directions = [ 'rtl', 'ltr' ];
+
+const directionMixin = {
+	watch: {
+		direction: ( newValue: string ) => {
+			window.document.documentElement.setAttribute( 'dir', newValue );
+		}
+	},
+	beforeMount() {
+		window.document.documentElement.dir = 'ltr';
+	}
+};
+
 export default {
 	title: 'Components/Input',
 	component: WvuiInput,
@@ -15,13 +28,18 @@ export default {
 export const configurable = (): Vue.Component =>
 	Vue.extend( {
 		components: { WvuiInput },
+		mixins: [ directionMixin ],
 		props: {
 			disabled: { type: Boolean, default: boolean( 'Disabled', false ) },
 			type: {
 				type: String as PropType<keyof typeof InputType>,
 				default: select( 'Input Type', Object.keys( InputType ), 'Search' )
 			},
-			placeholder: { type: String, default: text( 'Placeholder', 'Search…' ) }
+			placeholder: { type: String, default: text( 'Placeholder', 'Search…' ) },
+			direction: {
+				type: String,
+				default: select( 'Lang Direction', directions, 'ltr' )
+			}
 		},
 		data() {
 			return {
@@ -52,8 +70,13 @@ export const configurable = (): Vue.Component =>
 export const withStartIcon = (): Vue.Component =>
 	Vue.extend( {
 		components: { WvuiInput },
+		mixins: [ directionMixin ],
 		props: {
-			disabled: { type: Boolean, default: boolean( 'Disabled', false ) }
+			disabled: { type: Boolean, default: boolean( 'Disabled', false ) },
+			direction: {
+				type: String,
+				default: select( 'Lang Direction', directions, 'ltr' )
+			}
 		},
 		data() {
 			return {
@@ -76,8 +99,13 @@ export const withStartIcon = (): Vue.Component =>
 export const withEndIcon = (): Vue.Component =>
 	Vue.extend( {
 		components: { WvuiInput },
+		mixins: [ directionMixin ],
 		props: {
-			disabled: { type: Boolean, default: boolean( 'Disabled', false ) }
+			disabled: { type: Boolean, default: boolean( 'Disabled', false ) },
+			direction: {
+				type: String,
+				default: select( 'Lang Direction', directions, 'ltr' )
+			}
 		},
 		data() {
 			return {
@@ -99,8 +127,13 @@ export const withEndIcon = (): Vue.Component =>
 export const withClearAction = (): Vue.Component =>
 	Vue.extend( {
 		components: { WvuiInput },
+		mixins: [ directionMixin ],
 		props: {
-			disabled: { type: Boolean, default: boolean( 'Disabled', false ) }
+			disabled: { type: Boolean, default: boolean( 'Disabled', false ) },
+			direction: {
+				type: String,
+				default: select( 'Lang Direction', directions, 'ltr' )
+			}
 		},
 		data() {
 			return {
@@ -127,8 +160,13 @@ export const withClearAction = (): Vue.Component =>
 export const withButton = (): Vue.Component =>
 	Vue.extend( {
 		components: { WvuiInput, WvuiButton },
+		mixins: [ directionMixin ],
 		props: {
-			disabled: { type: Boolean, default: boolean( 'Disabled', false ) }
+			disabled: { type: Boolean, default: boolean( 'Disabled', false ) },
+			direction: {
+				type: String,
+				default: select( 'Lang Direction', directions, 'ltr' )
+			}
 		},
 		template: `
 		<div class="sb-input sb-input--has-button">
@@ -148,8 +186,11 @@ const searchLanguageMap = {
 	Japanese: '探す',
 	Greek: 'Αναζήτηση',
 	Swedish: 'Söka',
-	Mazandeerani: 'جستجو کردن'
+	Hebrew: 'לחפש',
+	Arabic: 'بحث'
 };
+
+const rtlWords = [ 'לחפש', 'بحث' ];
 
 export const wikipediaSearchInput = (): Vue.Component =>
 	Vue.extend( {
@@ -161,11 +202,22 @@ export const wikipediaSearchInput = (): Vue.Component =>
 				default: select( 'Label language', searchLanguageMap, 'Search' )
 			}
 		},
+		computed: {
+			placeholder() {
+				return `${this.buttonLabel}…`;
+			}
+		},
+		watch: {
+			buttonLabel( newValue ) {
+				const isRtl = rtlWords.indexOf( newValue ) >= 0;
+
+				document.documentElement.dir = isRtl ? 'rtl' : 'ltr';
+			}
+		},
 		template: `
 		<div class="sb-input sb-input--has-button">
 			<wvui-input
-				placeholder="Search…"
-				icon="search"
+				:placeholder="placeholder"
 				:disabled="disabled"
 				start-icon="test"
 				:clearable="true"